requirements.txt
pygame
googlesearch-python
qrcode
pyautogui
pyttsx3
winapps
matplotlib
tkcalendar
pyqt5
pyqtwebengine
yfinance
pillow
openpyxl
sympy
pytube
hijri_converter
requests
instaloader
opencv-python
textblob
vaderSentiment
langdetect
screen-brightness-control
numpy
prophet
seaborn
mediapipe
pyshorteners

